Added bit-banging option
Moved all P2OUT, P2IN type
statements to the header so they can be changed
globally. Modified the driver to switch between bit banging
and USCI. (Note: The USCI code in place now is for the MSP430g2553 and
is not functional. This could be a Silicone problem with the MCU)